refactor: 프로필 수정 리팩토링
