+++ src/pages/MyPage/ChangeProfile/index.tsx
@@ -1,9 +1,7 @@
 import { FC, useState, useRef } from 'react';
-import { useNavigate } from 'react-router-dom';
 import { Formik, Field } from 'formik';
 import CustomButton from '../../../components/Common/CustomButton';
 import Layout from '../../../components/Common/Layout';
-import { changeProfile } from '../../../api/User';
 import { buttonStyle } from '../../../styles/customChakraPropsStyle';
 import { Header } from '../../../components/Common';
 import { useMe } from '../../../stores/meStore';
@@ -22,23 +20,19 @@ import {
   Button,
   useDisclosure,
 } from '@chakra-ui/react';
-import { 잘못된요청, 에러404, 서버오류 } from '../../../constant/constant';
 import WithdrawalModal from '../WithdrawalModal';
-import useErrorToast from '../../../hooks/useErrorToast';
-import useSuccessToast from '../../../hooks/useSuccessToast';
+import useChangeProfileMutation from '../../../queries/useChangeProfileMutation';
 
 const ChangeProfile: FC = () => {
-  const navigate = useNavigate();
-  const { me, updateMe } = useMe();
+  const { me } = useMe();
   const [imgFile, setImgFile] = useState<string>(me.profile_image_url);
   const upload = useRef<HTMLInputElement | null>(null);
   const [file, setFile] = useState<File | null>(null);
-
   const { isOpen, onOpen, onClose } = useDisclosure();
   const initialRef = useRef(null);
   const finalRef = useRef(null);
-  const { errorToast } = useErrorToast();
-  const { successToast } = useSuccessToast();
+  const mutation = useChangeProfileMutation(imgFile);
+
   const imgUpload = () => {
     if (upload.current?.files) {
       const file = upload.current.files[0];
@@ -70,31 +64,7 @@ const ChangeProfile: FC = () => {
     if (file) {
       formData.append('profileImage', file);
     }
-
-    try {
-      const response = await changeProfile(formData);
-      if (response.status === 200) {
-        updateMe({
-          ...values,
-          profile_image_url: imgFile,
-        });
-        successToast({
-          title: '프로필 변경 성공!',
-          description: `프로필을 변경하였습니다.`,
-        });
-        navigate('/');
-      }
-    } catch (err: any) {
-      if (err.response.status === 400) {
-        errorToast(잘못된요청);
-      }
-      if (err.response.status === 404) {
-        errorToast(에러404);
-      }
-      if (err.response.status === 500) {
-        errorToast(서버오류);
-      }
-    }
+    mutation.mutate(formData);
   };
 
   return (
